finish rewriting js to ts
