+++ src/backend/masterState.ts
@@ -3,22 +3,17 @@ import 'core-js';
 /* eslint-disable no-plusplus */
 /* eslint-disable guard-for-in */
 /* eslint-disable no-restricted-syntax */
-const componentActionsRecord = {};
-let index = 0;
+const componentActionsRecord : Array<any> = [];
+let index : number = 0;
 
 export default {
-  saveNew: (state, component) => {
+  saveNew: (state, component) : number => {
     componentActionsRecord[index] = { state, component };
     index++;
     return index - 1;
   },
-  getRecordByIndex: inputIndex => componentActionsRecord[inputIndex],
-  getComponentByIndex: inputIndex => (componentActionsRecord[inputIndex]
+  getRecordByIndex: (inputIndex : number) : any => componentActionsRecord[inputIndex],
+  getComponentByIndex: (inputIndex: number) : any => (componentActionsRecord[inputIndex]
     ? componentActionsRecord[inputIndex].component
     : undefined),
 };
-
-class twoHaha {
-  private componenActionsRecord : Array<any>;
-  private index = 0; 
-}
